Extract GenericPeripheral to a dedicated file
