+++ example/lib/adapter/ble_adapter.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 
+import 'package:blemulator_example/example_peripherals/generic_peripheral.dart';
 import 'package:blemulator_example/model/ble_peripheral.dart';
 import 'package:blemulator_example/example_peripherals/sensor_tag.dart';
 import 'package:flutter_ble_lib/flutter_ble_lib.dart';
